Move basic pages from routes to pages
