+++ frontend/src/routes/index.tsx
@@ -2,7 +2,7 @@ import { Routes, Route, Navigate } from "react-router-dom";
 import AuthRoute from "../components/auth/AuthRoute";
 import PageContainer from "../components/layout/PageContainer";
 import { UserRole } from "../types/global";
-import AboutPage from "./AboutPage";
+import AboutPage from "../pages/AboutPage";
 import CallApplicationsPage from "../pages/CallApplicationsPage";
 import CallDetailPage from "../pages/CallDetailPage";
 import CallManagementPage from "../pages/CallManagementPage";
@@ -10,11 +10,11 @@ import CallPreviewPage from "../pages/CallPreviewPage";
 import CallsPage from "../pages/CallsPage";
 import DashboardPage from "../pages/DashboardPage";
 import LoginPage from "../pages/LoginPage";
-import MyApplicationsPage from "./MyApplicationsPage";
-import NotFoundPage from "./NotFoundPage";
+import MyApplicationsPage from "../pages/MyApplicationsPage";
+import NotFoundPage from "../pages/NotFoundPage";
 import RegisterPage from "../pages/RegisterPage";
 import ReviewPage from "../pages/ReviewPage";
-import ReviewerPage from "./ReviewerPage";
+import ReviewerPage from "../pages/ReviewerPage";
 import ApplicationLayout from "../pages/calls/apply/ApplicationLayout";
 import Step1_CallInfo from "../pages/calls/apply/Step1_CallInfo";
 import Step2_Upload from "../pages/calls/apply/Step2_Upload";
